Update I2C drivers to support read/write operations over 255 bytes (16-bit length field)

+++ libraries/AP_HAL/I2CDriver.h
@@ -14,15 +14,15 @@ public:
     virtual void setHighSpeed(bool active) = 0;
 
     /* write: for i2c devices which do not obey register conventions */
-    virtual uint8_t write(uint8_t addr, uint8_t len, uint8_t* data) = 0;
+    virtual uint8_t write(uint8_t addr, uint16_t len, uint8_t* data) = 0;
     /* writeRegister: write a single 8-bit value to a register */
     virtual uint8_t writeRegister(uint8_t addr, uint8_t reg, uint8_t val) = 0;
     /* writeRegisters: write bytes to contigious registers */
     virtual uint8_t writeRegisters(uint8_t addr, uint8_t reg,
-                                   uint8_t len, uint8_t* data) = 0;
+                                   uint16_t len, uint8_t* data) = 0;
 
     /* read: for i2c devices which do not obey register conventions */
-    virtual uint8_t read(uint8_t addr, uint8_t len, uint8_t* data) = 0;
+    virtual uint8_t read(uint8_t addr, uint16_t len, uint8_t* data) = 0;
     /* readRegister: read from a device register - writes the register,
      * then reads back an 8-bit value. */
     virtual uint8_t readRegister(uint8_t addr, uint8_t reg, uint8_t* data) = 0;
@@ -30,13 +30,13 @@ public:
     /* readRegisters: read contigious device registers - writes the first 
      * register, then reads back multiple bytes */
     virtual uint8_t readRegisters(uint8_t addr, uint8_t reg,
-                                  uint8_t len, uint8_t* data) = 0;
+                                  uint16_t len, uint8_t* data) = 0;
 
 #if CONFIG_HAL_BOARD == HAL_BOARD_LINUX
     /* readRegistersMultiple: read contigious device registers. 
        Equivalent to count calls to readRegisters() */
     virtual uint8_t readRegistersMultiple(uint8_t addr, uint8_t reg,
-                                          uint8_t len, uint8_t count, 
+                                          uint16_t len, uint8_t count, 
                                           uint8_t* data) = 0;
 #endif
